+++ business-rules/responders/sprint/closesprint.go
@@ -5,14 +5,14 @@ import "github.com/moul/cleanarch"
 type CloseSprintResponse interface {
 	cleanarch.UseCaseResponse
 
-	GetAverageClosedIssues() int
+	GetAverageClosedIssues() float64
 	GetClosedIssuesCount() int
 	GetSprintID() int
 }
 
 type CloseSprintResponseBuilder interface {
 	Create() CloseSprintResponseBuilder
-	WithAverageClosedIssues(int) CloseSprintResponseBuilder
+	WithAverageClosedIssues(float64) CloseSprintResponseBuilder
 	WithClosedIssuesCount(int) CloseSprintResponseBuilder
 	WithSprintID(int) CloseSprintResponseBuilder
 	Build() (CloseSprintResponse, error)
